React-App-Scene/flask-backend/requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==23.2.0
black==24.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dataclasses==0.6
defusedxml==0.7.1
filelock==3.13.3
fire==0.6.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
imantics==0.1.12
imgaug==0.4.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.1.1
kiwisolver==1.4.5
labelme2coco==0.2.4
lazy_loader==0.3
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.17.1
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pillow==10.3.0
pixellib==0.7.1
platformdirs==4.2.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pybboxes==0.1.6
pycocotools==2.0.7
pydot==2.0.0
Pygments==2.17.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
sahi==0.11.15
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
shapely==2.0.3
six==1.16.0
supervision==0.19.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.45
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
xmljson==0.2.1
yacs==0.1.8
